% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_Arrs.R
\name{spl_Arrs}
\alias{spl_Arrs}
\title{Arrival time generation for a cubic spline function composed of cardinal B-splines}
\usage{
spl_Arrs(knots, c, m)
}
\arguments{
\item{knots}{The knots upon which the B-spline is built - output 2 of the spl_Fit function (spl_Fit$Knots)}

\item{c}{The spline-coefficients of the spline function - output 1 of the spl_Fit function (spl_Fit$Opt_Spl_Coeffs)}

\item{m}{The number of days of arrivals required}
}
\value{
Arrivals - An unsorted vector containing the arrivals from day 1 to m one day after another
}
\description{
This thinning-based algorithm returns arrivals from the cubic spline function arrival rate
}
\examples{
## spline-based intensity defined on [0,10]
d<-3        #spline function degree
knots<-seq(-3,13,by=1)
n<-length(knots)-(d+1)
c<-runif(n,1,5)
m<-5
## plot the spline-based function and the arrivals
x<-seq(0,10-0.01,by=0.01)
plot(x,spl(x,knots,c),lwd=0.5,ylim=c(0,5),type="l")
Arrs<-spl_Arrs(knots,c,m)
points(sort(Arrs),rep(0,length(Arrs)))
}
